Initial version of WpfMusicPlayer
Application properly loads music.json as a local asset using Newtonsoft.Json package. 
Model, Repository, ViewModel and View classes are in place.
MVVM pattern takes advantage of Prism packages to bind View elements to ViewModel.

UI is limited to display name of artists, name of albums for the selected artists and name of songs for selected album in separate listviews.
Intent was to prove proper data binding first, before extending the UI to more elaborate views for albums and songs.
Application is coded to "happy" case at this point and data is loaded synchronously.

Next planned step is to add user controls to provide functionality needed for albums and songs listviews.
Sorting is not yet properly handled.
Application has a few defects such as the update on number of albums and songs.